requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
sqlalchemy
psycopg2-binary
pytesseract
pillow
dateparser
spacy
pdf2image
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pytest
pytest-asyncio
